requirements.txt
openai
uvicorn
fastapi
python-multipart
loguru
deepgram-sdk
websocket
nest-asyncio